Added more teleporters and notes
Learned how to read a map and moved teleport to dungeon 2 to the right place but couldn't figure out where dungeon 1 was supposed to be accessed from. Also added temp teleporter from plains1 to plains2 bc the original idea was to let the player look over there and not be able to go across the river.